+++ src/components/index.js
@@ -1,4 +1,3 @@
 export { default as SmarterChat } from "./SmarterChat";
-export { default as Console } from "./Console";
+export { Console, MenuItems } from "./Console";
 export { MessageDirectionEnum, SenderRoleEnum, ValidMessageRolesEnum } from "./enums";
-export { default as version } from "../version";