+++ src/index.js
@@ -1,0 +1,4 @@
+
+export { SmarterChat, Console, MessageDirectionEnum, SenderRoleEnum, ValidMessageRolesEnum, MenuItems } from "./components";
+export { ConfigPropTypes } from "./types";
+export { version } from "./version";