+++ src/types/index.js
@@ -1,0 +1,2 @@
+import ConfigPropTypes from "./propTypes";
+export { ConfigPropTypes };