+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.1
+
+- integrate Console to SmarterChat as include by default.
+- create a PropTypes definition for config object
+- validate config object in SmarterChat fetch
+- export all enumerated data types, Console, and PropTypes
+
 ## 0.2.0
 
 - add server console output.